Wired the logic for saving optimisation
Wired the logic for optimisation save to database with the logic for requesting optimisation run from the optimiser. Set up a data mapper, that maps the differnet models to whatever type of a model is needed. The call for saving to database is called in the create controller, after optimisation is received.

+++ services/rdm/Rdm.Api/Application/Model/OptimisationResultsHourly.cs
@@ -3,10 +3,10 @@
 public class OptimisationResultsHourly
 {
     public int Id { get; set; }
-    public float HeatProduction { get; set; }
-    public float ElectricityConsumption { get; set; }
-    public float Co2Emissions { get; set; }
-    public float Expenses { get; set; }
+    public double HeatProduction { get; set; }
+    public double ElectricityConsumption { get; set; }
+    public double Co2Emissions { get; set; }
+    public double Expenses { get; set; }
     public DateTime TimeFrom { get; set; }
     public DateTime TimeTo { get; set; }
     public List<ProductionUnit> ProductionUnits { get; set; }
